requirements.txt
numpy 
transformers
tensorboardX
multiprocess 
pyrouge
pytest
pytorch_pretrained_bert 